Leetcode daily #474 Ones and Zeroes
A 3D dp problem. Wow. That was something else.
An extension of the coin change problem ofc, but with quite a lot of factors combined. I think I have got a good hand on how to
memoize any given recursive solution that I come up with. But I need more practice with the Tabulation part.
Here the prime trick is coming up with the StringFrequency array and following that everything is the same as the CoinChange problem with the exception of the Max instead of finding the minimum.